% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-compare.R
\name{plot.aucmat_compare}
\alias{plot.aucmat_compare}
\title{Forest plot of paired AUC differences}
\usage{
\method{plot}{aucmat_compare}(x, ...)
}
\arguments{
\item{x}{An \code{aucmat_compare} data.frame.}

\item{...}{Ignored.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plots pairwise AUC differences with confidence intervals from a
\code{\link[=compare_auc]{compare_auc()}} result.
}
