% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_auc_rank}
\alias{plot_auc_rank}
\title{Rank plot: ordered discrimination strengths}
\usage{
plot_auc_rank(fit, n_label = 20L, show_ci = TRUE)
}
\arguments{
\item{fit}{An \code{aucmat_screen} object.}

\item{n_label}{Number of top biomarkers to label.  Default 20.}

\item{show_ci}{If \code{TRUE} and CIs are present, add error bars.}
}
\value{
A \code{ggplot2} object.
}
\description{
Rank plot: ordered discrimination strengths
}
