% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{scale_colour_aucmat_significance}
\alias{scale_colour_aucmat_significance}
\title{aucmat significance scale}
\usage{
scale_colour_aucmat_significance(q_cutoff = 0.05, ...)
}
\arguments{
\item{q_cutoff}{Cutoff value for the legend label.}

\item{...}{Passed to \code{\link[ggplot2:scale_manual]{ggplot2::scale_colour_manual()}}.}
}
\description{
Red for significant, grey for non-significant.
}
\keyword{internal}
