% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{scale_fill_aucmat_direction}
\alias{scale_fill_aucmat_direction}
\title{aucmat fill colour scale (discrete)}
\usage{
scale_fill_aucmat_direction(...)
}
\arguments{
\item{...}{Passed to \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.}
}
\description{
aucmat fill colour scale (discrete)
}
