% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_analytical.R
\name{generate_data_analytical}
\alias{generate_data_analytical}
\title{Generate Synthetic Data with Specific AUC and Correlation}
\usage{
generate_data_analytical(n, prevalence, target_aucs, corr_matrix)
}
\arguments{
\item{n}{number of rows (observations)}

\item{prevalence}{the proportion of the positive class for the binary outcome}

\item{target_aucs}{a vector of p target AUC values}

\item{corr_matrix}{a p x p target correlation matrix}
}
\description{
This function analytically generates a dataset with specified properties.
}
\examples{
\donttest{
set.seed(42)
sim <- generate_data_analytical(n=100, prevalence=0.3,
  target_aucs=c(0.8,0.7), corr_matrix=matrix(c(1,0.3,0.3,1),2,2))
round(sim$achieved_aucs, 3)
}
}
