% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-simulation.R
\name{plot.aucmat_simulation}
\alias{plot.aucmat_simulation}
\title{Plot simulated biomarker data}
\usage{
\method{plot}{aucmat_simulation}(x, ...)
}
\arguments{
\item{x}{An \code{aucmat_simulation} object.}

\item{...}{Ignored.}
}
\value{
A \code{ggplot2} object (via GGally if available) or a base-R pairs plot.
}
\description{
Default plot for \code{aucmat_simulation} objects. Shows a pairs plot of
biomarker columns with class-coloured density panels.
}
