% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_auc_stability}
\alias{plot_auc_stability}
\title{Stability plot: rank distributions and top-k probabilities}
\usage{
plot_auc_stability(stability, n_label = 25L)
}
\arguments{
\item{stability}{An \code{aucmat_stability} object.}

\item{n_label}{Number of biomarkers to show.  Default 25.}
}
\value{
A \code{ggplot2} object.
}
\description{
Stability plot: rank distributions and top-k probabilities
}
